Update dependency paths for files moved into the src directory.
