+++ jobs/rotate_featured.js
@@ -1,7 +1,7 @@
-const fq = require('../util/featuredQueue');
-const FeaturedQueue = require('../dynamo/models/featuredQueue');
-const util = require('../util/util');
-const User = require('../dynamo/models/user');
+const fq = require('../src/util/featuredQueue');
+const FeaturedQueue = require('../src/dynamo/models/featuredQueue');
+const util = require('../src/util/util');
+const User = require('../src/dynamo/models/user');
 
 (async () => {
   try {
